requirements.txt
psycopg2-binary==2.9.9
boto3>=1.28.0
duckdb>=0.9.0
rapidfuzz>=3.0.0
requests>=2.31.0
pandas>=2.0.0
gender-guesser>=0.4.0
